AI黑客松SAGE创新性维度评估/AI黑客松SAGE创新性维度评估/requirements.txt
# GitHub API
PyGithub>=2.1.1

# 魔搭社区
modelscope>=1.9.0

# Web界面
gradio>=4.0.0

# 限流重试
tenacity>=8.2.0

# 缓存
diskcache>=5.6.0

# 环境变量
python-dotenv>=1.0.0

# 数据处理
numpy>=1.24.0
requests>=2.28.0
